pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meta-agent"
version = "0.1.0"
description = "Meta-Agent: An LLM System that Designs, Builds, and Orchestrates Other LLM Agents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Meta-Agent Team"},
]
keywords = ["llm", "agents", "orchestration", "meta-agent", "ai-systems"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core LLM & Agent frameworks
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",

    # API layer
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "python-multipart>=0.0.12",

    # Data validation & serialization
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Memory & storage
    "redis>=5.2.0",
    "chromadb>=0.5.0",

    # Observability
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",

    # Utilities
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "networkx>=3.4.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.8.0",
    "jinja2>=3.1.4",
    "ddgs>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

benchmarks = [
    "mlflow>=2.18.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
]

[project.scripts]
meta-agent = "meta_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
